requirements.txt
tweepy
python-dotenv
requests
groq
Pillow
